requirements.txt
# Core dependencies
beautifulsoup4>=4.13.0
certifi>=2024.0.0
customtkinter>=5.2.0
darkdetect>=0.8.0
pillow>=10.0.0
psutil>=6.0.0
requests>=2.31.0
Send2Trash>=1.8.0
tqdm>=4.66.0
ttkbootstrap>=1.13.0

# Archive support
py7zr>=1.0.0
pybcj>=1.0.0
inflate64>=1.0.0
multivolumefile>=0.2.0
pyppmd>=1.2.0
pyzstd>=0.16.0
Brotli>=1.1.0
pycryptodomex>=3.20.0

# Networking
PySocks>=1.7.0
urllib3>=2.0.0
websocket-client>=1.0.0

# Configuration
python-dotenv>=1.0.0

# Windows-only dependencies
pywin32-ctypes>=0.2.0; sys_platform == 'win32'
pefile>=2023.0.0; sys_platform == 'win32'
colorama>=0.4.0; sys_platform == 'win32'
